controller/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-json-logger==2.0.7
kubernetes==30.1.0
httpx==0.27.2
PyYAML==6.0.2
prometheus-client==0.21.0
sqlalchemy
psycopg2-binary
